Correct way of processing translations
- Use source dir as input. This way *.ui files are scanned and used
directly, rather than their generated header files. This way locations
are much more stable (always in the same file - not from the build
directory - and not in some random position within the file).
- Properly use "en_US" as source language (not "en-US")

Note: there are no actual translation changes in this commit. The only
changes made is where translations had hotkeys added to them in the
source language (because of using the *.ui files). These have been
fixed - but the "old" vanished strings are still there for now.