Select best available resolutions
Instead of forcing a specific item (based on the text), the Sims 2 +
Stories plugins now attempt to find the best resolution when loading.
This is either the exact match if available, or a lower resolution.

This specifically solves the problem when one of the default settings'
resolutions is not available on the PC, which would lead to a fallback
to the lowest available resolution.

Issue #8