Rewrite FindDirectX9 based on Ogre3d.com DirectX11 version
This picks up the Windows SDK version if available, over the legacy
DirectX9 SDK. They cannot be used in combination, but will conflict if
the headers from the older one is included first.

This is based on
https://raw.githubusercontent.com/OGRECave/ogre/master/CMake/Packages/FindDirectX11.cmake
with these changes:
- Change DirectX11 to DirectX9
- Explicitly define DirectX9_DXGI_LIBRARY (and a few others) if just the
include dir is found (from the Windows SDK).

This all seems to build more reliably on newer systems