Delegate plugin settings to GraphicsRulesWriter
As this class also actually loads the plugin, it can much better also
handle loading settings after loading the plugin - including the path -
and then also saving them (for symmetry).

The MainWindow is now more concerned with replacing the actual widget
and a bit of control, but doesn't deal with the registry settings
directly anymore.

All of this leads to more reliable (and easier to understand) behavior
on all settings.