Also trigger Manual Save behavior if writing of files fails
This may happen if the back-ups already exist (so that passes through
fine and doesn't trigger it), but overwriting the files still fails
because of lack of permissions.

This will now also ask the user to copy the files manually (from a
temporary location), instead of them getting just a useless error
message.