More robust installation of qwindows platform plugin
The MSVC version may require qwindowsd.dll instead of qwindows.dll (in a
debug environment). This uses the proper imported location for each
configuration, if it is defined. Otherwise it falls back to the default
path and name.

It uses IMPORT_CONFIGURATIONS to determine if the import is correctly
done (not so on the Mingw version!). Using the LOCATION directly results
in CMake errors causing the configuration or generation to fail. This
works safely and doesn't need any policies to be set to OLD.