Disable texture memory adjustment on all cards (Vista+) by default
This changes the default so that all modern drivers (Windows Vista and
later) are expected to report accurate values for texture memory.

This option seems to be work well for all cards and vendors, so no need
to keep it Nvidia only by default anymore.

As the Windows driver model changed from XP to Vista, on Windows XP the
tweak is still disabled by default.

Also reorganized the code a little bit to take out an #ifdef Windows.