Enable adding of new video card vendors
Based on the PCI ID repository at http://pci-ids.ucw.cz/ - this takes
the pci.ids file, discards all non-vendor lines (most of them) and
writes a header file of vendors with their ids and names.

This header is then used to provide the name of the vendor when it needs
to be added. This allows adding vendors to a Video Cards database, even
if the database file doesn't contain the vendor yet.

This somewhat round-about way is necessary as Windows doesn't seem to
provide an easy to use API to get the vendor name from the graphics
device.

Because the original PCI IDs list is quite a big file, the limited
header is generated during the build instead of reading the file at
runtime. This reduces the executable's file size and improves runtime
performance.